Portfolio Part 1
Links for major projects added with the exception of the battle arena.
Styling still needs work. Each project will be in a box similar to
linked in with an image header, icons for github and a web-globe, and
hyperlinks to github and website as appropriate. Still need pictures of
bee-pesticide-model, battle-arena, and bird database. Afterward, just
design your contact section and add the javascript. :)