Animation Working
Animation now works as intended. React code is significantly cleaned up
too.

1. Iframes have been abandoned thanks to x-header-options blocking me.
It's not worth it to me to try and get around that when it's easier just
to take up to date screenshots.
2. Flex-shrink: 0 is a nice trick to crop images with 92% coverage. :)
3. Warframe's style is no longer on the table like I originally planned.
The expansion will take way too long to figure out programmatically
because of positioning on mobile and the need for responsive design.
Using absolute positiong would be a nightmare, especially for mobile.
Therefore, I'm going to focus on just an animated slideshow with fade-in
links for modern browsers and a graceful degradation to a static
slideshow for older browsers.