React Portfolio 0.4
Mostly done except for 2 things:
1. Add project details in proper.
2. Fix bug with styles of "current-slide" when in mobile(has a gray
padding line above it when it shouldn't)

Fade-in abandoned because of the way the slideshow styles are done. It
would look really ugly if I did a fade-in, plus I feel that speed is
more valuable than effects at this point. The animation would be too
long if users had to wait for the image and then wait longer for the
details. For now, I'll just animate the image for flair.